# docs/openapi.yaml
openapi: 3.1.0
info:
  title: edu-platform
  version: 0.1.0
paths:
...
  /v1/leaderboard:
    get:
      tags:
      - ranking
      summary: Leaderboard
      operationId: leaderboard_v1_leaderboard_get
      parameters:
      - name: scope
        in: query
        required: false
        schema:
          type: string
          default: total
          title: Scope
      - name: key
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Key
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          default: 50
          title: Limit
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LeaderboardOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/me:
    get:
      tags:
      - ranking
      summary: Me Overview
      operationId: me_overview_v1_me_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/me/analysis:
    get:
      tags:
      - ranking
      summary: My Analysis
      operationId: my_analysis_v1_me_analysis_get
      parameters:
      - name: accept-language
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Accept-Language
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/parent/link-code:
    post:
      tags:
      - parent
      summary: Create Link Code
      operationId: create_link_code_v1_parent_link_code_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LinkCodeOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/parent/link:
    post:
      tags:
      - parent
      summary: Link Child
      operationId: link_child_v1_parent_link_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LinkIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/parent/children:
    get:
      tags:
      - parent
      summary: List Children
      operationId: list_children_v1_parent_children_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChildrenOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/parent/children/{student_id}:
    get:
      tags:
      - parent
      summary: Child Detail
      operationId: child_detail_v1_parent_children__student_id__get
      parameters:
      - name: student_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Student Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChildSummary'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/parent/children/{student_id}/analysis:
    get:
      tags:
      - parent
      summary: Child Analysis
      operationId: child_analysis_v1_parent_children__student_id__analysis_get
      parameters:
      - name: student_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Student Id
      - name: accept-language
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Accept-Language
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ChildSummary:
      properties:
        student_id:
          type: string
          title: Student Id
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        grade:
          anyOf:
          - type: integer
          - type: 'null'
          title: Grade
        region_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Region Code
        progress:
          $ref: '#/components/schemas/ProgressOut'
      type: object
      required:
      - student_id
      - progress
      title: ChildSummary
    ChildrenOut:
      properties:
        children:
          items:
            $ref: '#/components/schemas/ChildSummary'
          type: array
          title: Children
          default: []
      type: object
      title: ChildrenOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    LeaderboardEntry:
      properties:
        rank:
          type: integer
          title: Rank
        user_id:
          type: string
          title: User Id
        display_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Display Name
        region_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Region Code
        score:
          type: integer
          title: Score
        is_me:
          type: boolean
          title: Is Me
          default: false
      type: object
      required:
      - rank
      - user_id
      - score
      title: LeaderboardEntry
    LeaderboardOut:
      properties:
        scope:
          type: string
          title: Scope
        key:
          anyOf:
          - type: string
          - type: 'null'
          title: Key
        entries:
          items:
            $ref: '#/components/schemas/LeaderboardEntry'
          type: array
          title: Entries
        me:
          anyOf:
          - $ref: '#/components/schemas/LeaderboardEntry'
          - type: 'null'
        total_ranked:
          type: integer
          title: Total Ranked
          default: 0
      type: object
      required:
      - scope
      title: LeaderboardOut
    LinkCodeOut:
      properties:
        code:
          type: string
          title: Code
        expires_in_seconds:
          type: integer
          title: Expires In Seconds
      type: object
      required:
      - code
      - expires_in_seconds
      title: LinkCodeOut
      description: 'A student generates this and reads it to a parent (or shares it).
        The

        parent enters it to establish guardianship. Short-lived, single-use.'
    LinkIn:
      properties:
        code:
          type: string
          title: Code
      type: object
      required:
      - code
      title: LinkIn
...
    ProgressOut:
      properties:
        xp:
          type: integer
          title: Xp
          default: 0
        level:
          type: integer
          title: Level
          default: 1
        streak_days:
          type: integer
          title: Streak Days
          default: 0
        answered:
          type: integer
          title: Answered
          default: 0
        correct:
          type: integer
          title: Correct
          default: 0
        accuracy:
          type: number
          title: Accuracy
          default: 0.0
      type: object
      title: ProgressOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError